+++ era-compiler-solidity/tests/common/const.rs
@@ -185,6 +185,22 @@ pub const TEST_JSON_CONTRACT_PATH_SUPPRESSED_WARNINGS_INVALID: &str =
     "tests/data/standard_json_input/suppressed_warnings_invalid.json";
 
 /// A test input file.
+pub const TEST_JSON_CONTRACT_PATH_METADATA_HASH_IPFS_AND_METADATA: &str =
+    "tests/data/standard_json_input/metadata_hash_ipfs_and_metadata.json";
+
+/// A test input file.
+pub const TEST_JSON_CONTRACT_PATH_METADATA_HASH_IPFS_NO_METADATA: &str =
+    "tests/data/standard_json_input/metadata_hash_ipfs_no_metadata.json";
+
+/// A test input file.
+pub const TEST_JSON_CONTRACT_PATH_METADATA_HASH_NONE_AND_METADATA: &str =
+    "tests/data/standard_json_input/metadata_hash_none_and_metadata.json";
+
+/// A test input file.
+pub const TEST_JSON_CONTRACT_PATH_METADATA_HASH_NONE_NO_METADATA: &str =
+    "tests/data/standard_json_input/metadata_hash_none_no_metadata.json";
+
+/// A test input file.
 pub const TEST_DISASSEMBLER_BYTECODE_PATH: &str = "tests/data/bytecodes/disassembler.zbin";
 
 /// A test input file.