+++ era-compiler-solidity/tests/unit/optimizer.rs
@@ -41,7 +41,7 @@ fn default(version: semver::Version, codegen: era_solc::StandardJsonInputCodegen
     let build_unoptimized = crate::common::build_solidity_standard_json(
         sources.clone(),
         era_compiler_common::Libraries::default(),
-        era_compiler_common::HashType::Keccak256,
+        era_compiler_common::EraVMMetadataHashType::Keccak256,
         BTreeSet::new(),
         &version,
         codegen,
@@ -51,7 +51,7 @@ fn default(version: semver::Version, codegen: era_solc::StandardJsonInputCodegen
     let build_optimized_for_cycles = crate::common::build_solidity_standard_json(
         sources.clone(),
         era_compiler_common::Libraries::default(),
-        era_compiler_common::HashType::Keccak256,
+        era_compiler_common::EraVMMetadataHashType::Keccak256,
         BTreeSet::new(),
         &version,
         codegen,
@@ -61,7 +61,7 @@ fn default(version: semver::Version, codegen: era_solc::StandardJsonInputCodegen
     let build_optimized_for_size = crate::common::build_solidity_standard_json(
         sources,
         era_compiler_common::Libraries::default(),
-        era_compiler_common::HashType::Keccak256,
+        era_compiler_common::EraVMMetadataHashType::Keccak256,
         BTreeSet::new(),
         &version,
         codegen,